Move barnard_exact to _hypotests (and the tests)
