+++ scipy/stats/tests/test_hypotests.py
@@ -4,11 +4,12 @@ import numpy as np
 from numpy.testing import (assert_, assert_equal, assert_allclose,
                            assert_almost_equal)  # avoid new uses
 
+import pytest
 from pytest import raises as assert_raises
 from scipy.stats._hypotests import (epps_singleton_2samp, cramervonmises,
                                     _cdf_cvm)
+from scipy.stats import distributions, barnard_exact
 import scipy.stats as stats
-from scipy.stats import distributions
 from .common_tests import check_named_results
 
 
@@ -405,3 +406,105 @@ class TestSomersD(object):
         y = np.arange(10)
         res = stats.somersd(x, y)
         assert_equal(res.table, np.eye(10))
+
+
+class TestBarnardExact(object):
+    """Some tests to show that barnard_exact() works correctly.
+    """
+
+    @pytest.mark.parametrize("input_sample,expected", [
+        ([[43, 40], [10, 39]], (-3.555406779643, 0.000362832367)),
+        ([[100, 2], [1000, 5]], (1.776382925679, 0.135126970878)),
+        ([[2, 7], [8, 2]], (2.518474945157, 0.019210815430)),
+        ([[5, 1], [10, 10]], (-1.449486150679, 0.156277546306)),
+        ([[5, 15], [20, 20]], (1.851640199545, 0.066363501421)),
+        ([[5, 16], [20, 25]], (1.609639949352, 0.116984852192)),
+        ([[10, 5], [10, 1]], (1.449486150679, 0.177536588915)),
+        ([[5, 0], [1, 4]], (-2.581988897472, 0.013671875000)),
+        ([[0, 1], [3, 2]], (1.095445115010, 0.509667991877)),
+        ([[0, 2], [6, 4]], (1.549193338483, 0.197019618792)),
+        ([[2, 7], [8, 2]], (2.518474945157, 0.019210815430)),
+    ])
+    def test_precise(self, input_sample, expected):
+        """The expected values have been generated by R, using a resolution
+        for the nuisance parameter of 1e-6.
+        """
+        res = barnard_exact(input_sample)
+        TX_obs, pvalue = res.score_statistic, res.pvalue
+        assert_allclose([TX_obs, pvalue], expected)
+
+    @pytest.mark.parametrize("input_sample,expected", [
+        ([[43, 40], [10, 39]], (-3.920362887717, 0.000289470662)),
+        ([[100, 2], [1000, 5]], (1.139432816087, 0.950272080594)),
+        ([[2, 7], [8, 2]], (3.079373904042, 0.020172119141)),
+        ([[5, 1], [10, 10]], (-1.622375939458, 0.150599922226)),
+        ([[5, 15], [20, 20]], (1.974771239528, 0.063038448651)),
+        ([[5, 16], [20, 25]], (1.722122973346, 0.133329494287)),
+        ([[10, 5], [10, 1]], (1.765469659009, 0.250566655215)),
+        ([[5, 0], [1, 4]], (-5.477225575052, 0.007812500000)),
+        ([[0, 1], [3, 2]], (1.224744871392, 0.509667991877)),
+        ([[0, 2], [6, 4]], (1.732050807569, 0.197019618792)),
+        ([[2, 7], [8, 2]], (3.079373904042, 0.020172119141)),
+    ])
+    def test_pooled_param(self, input_sample, expected):
+        """The expected values have been generated by R, using a resolution
+        for the nuisance parameter of 1e-6.
+        """
+        res = barnard_exact(input_sample, pooled=False)
+        TX_obs, pvalue = res.score_statistic, res.pvalue
+        assert_allclose([TX_obs, pvalue], expected)
+
+    def test_raises(self):
+        # test we raise an error for wrong input number of nuisances.
+        error_msg = (
+            "Number of iterations `num_it` must be non-negative,"
+            " found -1")
+        with assert_raises(ValueError, match=error_msg):
+            barnard_exact([[1, 2], [3, 4]], num_it=-1)
+
+        # test we raise an error for wrong shape of input.
+        error_msg = "The input `table` must be of shape \\(2, 2\\)."
+        with assert_raises(ValueError, match=error_msg):
+            barnard_exact(np.arange(6).reshape(2, 3))
+
+        # Test all values must be positives
+        error_msg = "All values in `table` must be nonnegative."
+        with assert_raises(ValueError, match=error_msg):
+            barnard_exact([[-1, 2], [3, 4]])
+
+        # Test value error on wrong alternative param
+        error_msg = (
+            "`alternative` should be one of {'two-sided', 'less', 'greater'},"
+            " found .*")
+        with assert_raises(ValueError, match=error_msg):
+            barnard_exact([[1, 2], [3, 4]], "not-correct")
+
+    @pytest.mark.parametrize("input_sample,expected", [
+        ([[0, 5], [0, 10]], (1.0, np.nan)),
+        ([[5, 0], [10, 0]], (1.0, np.nan)),
+    ])
+    def test_row_or_col_zero(self, input_sample, expected):
+        res = barnard_exact(input_sample)
+        TX_obs, pvalue = res.score_statistic, res.pvalue
+        assert_equal(pvalue, expected[0])
+        assert_equal(TX_obs, expected[1])
+
+    @pytest.mark.slow
+    @pytest.mark.parametrize("input_sample,expected", [
+        ([[2, 7], [8, 2]], (2.518474945157, 0.009886140845,
+        1 - 0.009886140845)),
+        ([[200, 7], [8, 300]], (-21.320036698460, 0.0, 1.0)),
+        ([[28, 21], [6, 1957]], (-30.489638143953, 0.0, 1.0)),
+    ])
+    def test_less_greater(self, input_sample, expected):
+        TX_expect, less_pvalue_expect, greater_value_expect = expected
+
+        res = barnard_exact(input_sample, alternative="less")
+        TX_obs, pvalue = res.score_statistic, res.pvalue
+        pvalue = min((pvalue, 1 - pvalue))
+        assert_allclose([TX_obs, pvalue], [TX_expect, less_pvalue_expect], atol=1e-7)
+
+        res = barnard_exact(input_sample, alternative="greater")
+        TX_obs, pvalue = res.score_statistic, res.pvalue
+        pvalue = max((pvalue, 1 - pvalue))
+        assert_allclose([TX_obs, pvalue], [TX_expect, greater_value_expect], atol=1e-7)
